# src/openapi/audition-api.swagger.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/posts": {
      "get": {
        "tags": [
          "audition-post-controller"
        ],
        "summary": "Fetch posts",
        "description": "Retrieve a list of audition posts. Optionally filter by user IDs.",
        "operationId": "getPosts",
        "parameters": [
          {
            "name": "userIds",
            "in": "query",
            "description": "List of user IDs to filter posts",
            "required": false,
            "schema": {
              "type": "array",
              "format": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of posts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditionPost"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user IDs supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditionPost"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/posts/{id}": {
      "get": {
        "tags": [
          "audition-post-controller"
        ],
        "summary": "Fetch post by ID",
        "description": "Retrieve an audition post by its ID.",
        "operationId": "getPostsById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the post to retrieve",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of the post",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditionPost"
                }
              }
            }
          },
          "404": {
            "description": "Post not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditionPost"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{id}/comments": {
      "get": {
        "tags": [
          "audition-post-controller"
        ],
        "summary": "Fetch comments for a post",
        "description": "Retrieve a list of comments for a specific audition post.",
        "operationId": "getCommentsByPostId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the post for which to fetch comments",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of comments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditionComment"
                }
              }
            }
          },
          "404": {
            "description": "Post not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditionComment"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/comments": {
      "get": {
        "tags": [
          "audition-comment-controller"
        ],
        "summary": "Fetch Comments by Post IDs",
        "description": "Retrieve a list of audition post comments. Optionally filter by post IDs.",
        "operationId": "getCommentsByPostIds",
        "parameters": [
          {
            "name": "postIds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved comments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditionPost"
                }
              }
            }
          },
          "400": {
            "description": "Invalid post IDs provided",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditionComment"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuditionPost": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          }
        }
      },
      "AuditionComment": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "body": {
            "type": "string"
          }
        }
      }
    }
  }
}